Moves DML AST nodes to dedicated package
