+++ partiql-ast/src/main/java/org/partiql/ast/AstVisitor.java
@@ -7,6 +7,20 @@ import org.partiql.ast.ddl.Ddl;
 import org.partiql.ast.ddl.KeyValue;
 import org.partiql.ast.ddl.PartitionBy;
 import org.partiql.ast.ddl.TableConstraint;
+import org.partiql.ast.dml.ConflictAction;
+import org.partiql.ast.dml.ConflictTarget;
+import org.partiql.ast.dml.Delete;
+import org.partiql.ast.dml.DoReplaceAction;
+import org.partiql.ast.dml.DoUpdateAction;
+import org.partiql.ast.dml.Insert;
+import org.partiql.ast.dml.InsertSource;
+import org.partiql.ast.dml.OnConflict;
+import org.partiql.ast.dml.Replace;
+import org.partiql.ast.dml.SetClause;
+import org.partiql.ast.dml.Update;
+import org.partiql.ast.dml.UpdateTarget;
+import org.partiql.ast.dml.UpdateTargetStep;
+import org.partiql.ast.dml.Upsert;
 import org.partiql.ast.expr.Expr;
 import org.partiql.ast.expr.ExprAnd;
 import org.partiql.ast.expr.ExprArray;
