+++ partiql-ast/src/main/java/org/partiql/ast/dml/Update.java
@@ -1,21 +1,26 @@
-package org.partiql.ast;
+package org.partiql.ast.dml;
 
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+import org.partiql.ast.AstNode;
+import org.partiql.ast.AstVisitor;
+import org.partiql.ast.IdentifierChain;
+import org.partiql.ast.Statement;
+import org.partiql.ast.expr.Expr;
 
 import java.util.ArrayList;
 import java.util.Collection;
 import java.util.List;
 
 /**
- * This is the insert statement.
- * @see InsertSource
+ * This is the update searched statement.
+ * @see SetClause
  */
 @Builder(builderClassName = "Builder")
 @EqualsAndHashCode(callSuper = false)
-public final class Insert extends Statement {
+public final class Update extends Statement {
     /**
      * TODO
      */
@@ -25,33 +30,25 @@ public final class Insert extends Statement {
     /**
      * TODO
      */
-    @Nullable
-    public final Identifier asAlias;
-
-    /**
-     * TODO
-     */
     @NotNull
-    public final InsertSource source;
+    public final List<SetClause> setClauses;
 
     /**
      * TODO
      */
     @Nullable
-    public final OnConflict onConflict;
+    public final Expr condition;
 
     /**
      * TODO
      * @param tableName TODO
-     * @param asAlias TODO
-     * @param source TODO
-     * @param onConflict TODO
+     * @param setClauses TODO
+     * @param condition TODO
      */
-    public Insert(@NotNull IdentifierChain tableName, @Nullable Identifier asAlias, @NotNull InsertSource source, @Nullable OnConflict onConflict) {
+    public Update(@NotNull IdentifierChain tableName, @NotNull List<SetClause> setClauses, @Nullable Expr condition) {
         this.tableName = tableName;
-        this.asAlias = asAlias;
-        this.source = source;
-        this.onConflict = onConflict;
+        this.setClauses = setClauses;
+        this.condition = condition;
     }
 
     @NotNull
@@ -59,11 +56,15 @@ public final class Insert extends Statement {
     public Collection<AstNode> children() {
         List<AstNode> kids = new ArrayList<>();
         kids.add(tableName);
+        kids.addAll(setClauses);
+        if (condition != null) {
+            kids.add(condition);
+        }
         return kids;
     }
 
     @Override
     public <R, C> R accept(@NotNull AstVisitor<R, C> visitor, C ctx) {
-        return visitor.visitInsert(this, ctx);
+        return visitor.visitUpdate(this, ctx);
     }
 }
