+++ partiql-ast/src/main/java/org/partiql/ast/dml/UpdateTarget.java
@@ -1,10 +1,11 @@
-package org.partiql.ast;
+package org.partiql.ast.dml;
 
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import org.jetbrains.annotations.NotNull;
-import org.jetbrains.annotations.Nullable;
-import org.partiql.ast.expr.Expr;
+import org.partiql.ast.AstNode;
+import org.partiql.ast.AstVisitor;
+import org.partiql.ast.Identifier;
 
 import java.util.ArrayList;
 import java.util.Collection;
