+++ partiql-ast/src/main/java/org/partiql/ast/dml/UpdateTargetStep.java
@@ -1,8 +1,11 @@
-package org.partiql.ast;
+package org.partiql.ast.dml;
 
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import org.jetbrains.annotations.NotNull;
+import org.partiql.ast.AstNode;
+import org.partiql.ast.AstVisitor;
+import org.partiql.ast.Identifier;
 import org.partiql.ast.expr.ExprLit;
 import org.partiql.value.PartiQL;
