+++ partiql-ast/src/main/java/org/partiql/ast/dml/Upsert.java
@@ -1,9 +1,14 @@
-package org.partiql.ast;
+package org.partiql.ast.dml;
 
 import lombok.Builder;
 import lombok.EqualsAndHashCode;
 import org.jetbrains.annotations.NotNull;
 import org.jetbrains.annotations.Nullable;
+import org.partiql.ast.AstNode;
+import org.partiql.ast.AstVisitor;
+import org.partiql.ast.Identifier;
+import org.partiql.ast.IdentifierChain;
+import org.partiql.ast.Statement;
 
 import java.util.ArrayList;
 import java.util.Collection;
