+++ partiql-ast/src/main/kotlin/org/partiql/ast/AstRewriter.kt
@@ -5,6 +5,20 @@ import org.partiql.ast.Ast.exprQuerySet
 import org.partiql.ast.Ast.identifier
 import org.partiql.ast.Ast.query
 import org.partiql.ast.ddl.Ddl
+import org.partiql.ast.dml.ConflictAction
+import org.partiql.ast.dml.ConflictTarget
+import org.partiql.ast.dml.Delete
+import org.partiql.ast.dml.DoReplaceAction
+import org.partiql.ast.dml.DoUpdateAction
+import org.partiql.ast.dml.Insert
+import org.partiql.ast.dml.InsertSource
+import org.partiql.ast.dml.OnConflict
+import org.partiql.ast.dml.Replace
+import org.partiql.ast.dml.SetClause
+import org.partiql.ast.dml.Update
+import org.partiql.ast.dml.UpdateTarget
+import org.partiql.ast.dml.UpdateTargetStep
+import org.partiql.ast.dml.Upsert
 import org.partiql.ast.expr.Expr
 import org.partiql.ast.expr.ExprAnd
 import org.partiql.ast.expr.ExprArray
